Vote cancelling is integrated into the vote form now
When there is an existing vote for a voteable and the corresponding vote
button is clicked the form will DELETE /votes/:vote_id.

Also add classes on buttons that have been voted on already and disables
the opposite button, too.

Much refactor needs to be done on the votes_helper still.