Voting now uses Ajax
Uses jQuery and jQuery.form. jQuery.form is needed because it serializes
the values of the submit button that is clicked, which allows us to know
whether the up or down vote button was pressed.

This also cleans up some of the views and creates a helper method,
html_id, which returns a string that can be used as an id to an HTML
element, like "answer-123".

Still need to handle unvoting (DELETE /votes/123) and errors.